+++ core/hostInfo/disk_windows.go
