+++ core/hostInfo/host_info_test.go
