+++ core/hostInfo/memory_unix.go
