+++ core/hostInfo/memory_windows.go
