+++ core/hostInfo/os_windows.go
